todos_app_workshop/requirements.txt
asgiref==3.8.0
Django==4.2.11
djangorestframework==3.15.0
sqlparse==0.4.4
typing_extensions==4.10.0
